requirements.txt
flask>=2.3.0
supabase-py>=2.0.0
python-dotenv>=1.0.0  # opcional, para .env local
